kernel: ralink_wifi: backport some fixes to MT7620 and MT7610 stable drivers.

+++ linux-3.4.x/drivers/net/wireless/ralink/rt3xxx/rt2860v2/common/cmm_mac_pci.c
@@ -30,7 +30,7 @@ VOID RTMPResetTxRxRingMemory(
 	TXD_STRUC       TxD;
 #endif /* RT_BIG_ENDIAN */
 	PNDIS_PACKET  pPacket;
-	unsigned int  IrqFlags;
+	ULONG IrqFlags;
 
 
 	/* Free TxSwQueue Packet*/
